Project_root/my-vue/requirements.txt
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
bcrypt==4.1.1
billiard==4.2.0
blinker==1.7.0
Brotli==1.1.0
cachelib==0.9.0
celery==5.3.5
cffi==1.16.0
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cssselect2==0.7.0
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Caching==2.1.0
Flask-Excel==0.0.7
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.45.0
greenlet==3.0.1
html5lib==1.1
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
lml==0.1.0
MarkupSafe==2.1.3
passlib==1.7.4
Pillow==10.1.0
prompt-toolkit==3.0.41
pycparser==2.21
pydyf==0.8.0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyparsing==3.1.1
pyphen==0.14.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
setuptools==69.0.2
six==1.16.0
SQLAlchemy==2.0.23
texttable==1.7.0
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.12
weasyprint==60.1
webencodings==0.5.1
Werkzeug==3.0.1
WTForms==3.1.1
zipp==3.17.0
zopfli==0.2.3
python-dotenv==1.0.1
